Updated Choropleth references
